Player movement, default vehicle setup.
Implemented the player movement and made sure that no matter the
position of the player's sprite, it is still being drawn by the Game
class. Setup the default vehicle sprite and texture and made sure it is
being drawn by the Game class. Laid out some template functions for the
random spawning of vehicles.

Cleaned up the current code and commented as needed.