Vehicle class changes.
Vehicle class now spawns a given number of sprites and assigns them
pre-defined textures. Also assigns them a direction and the position of
the sprite is changed to match the direction it has been given. None of
these values are currently random and the positional values are
currently only placeholder values.