Player movement changes and clean up.
Changed the player's movement to be static movement ie. pressing a
movement key will move the player in a direction by a constant value and
only once until key is released and pressed again. This will probably be
the method of movement from here on out as it will compliment the
physics soon to be implemented.

Cleaned up the current code and commented anything that needed
commenting.