Vehicle spawning and movement changes.
Fixed the bug where if more than 4 vehicles were on the top most road,
the final car to loop will get stuck on the first car going back around.
Also made it so the roads have a constant flow of vehicles. Finally,
made the spawning function a little more adaptable (bit more work todo
still) so that when more roads are added in the future, it should work
well, provided the correct positional values are given.