Collision, scoring and positional changes.
Got collision working for the player, vehicles and ending zone. The
player will lose a life (numOfLifes--) when hitting a vehicle and will
gain a point (score++) when hitting the ending zone. In both of those
situations the player is translated back to their starting position.

Changed Player.cpp and Vehicle.cpp constructor code into its own
Instantiation method so that values such as number of sprites, movement
speed, window dimensions etc can be passed through without having to
reconstruct the classes outside of the Game.h.

All objects now scale as per the window dimensions. Switched it to use
the percentage of the window (thanks mr Ryan Hood for the tip doot
doot).