UI, player sprite and stat changes.
Made sure both the score and the number of lives are adjusted on
collision with relevant objects.

Now show both the score and number of lives on screen during the game.

The frog sprite now faces the direction is is moving in.

note* bunch of useless/messy code atm whilst creating a main menu and
splitting into UI class.