UI changes.
UI is now controlled via a separate class and has various methods for
creating different kinds of UI. Makes the whole UI process more modular
and involves less repeating code. Unsure atm whether or not the way it
utilises the *window is the correct way to go about it, and may change
it.

Created the main menu and the player's HUD using the new UI system.