Window focus changes.
Window will now simulate the pausing of the game by setting the
deltaTime variable by which the cars are moved to 0 if the window is not
in focus and to the time since the last restarted clock cycle if it is
in focus. Every other event occurs via keyboard/mouse input and so
cannot be carried out without focus anyway.